+++ Woof/Woof/Model/KeyValueStorage/KeyValueStorage+Name.swift
@@ -5,5 +5,11 @@ extension KeyValueStorage {
     enum Name {
         /// The storage name to store information about current owner.
         static let currentOwner = "CurrentOwnerStorage"
+
+        /// The storage name to store information about current sitter.
+        static let currentSitter = "CurrentSitterStorage"
+
+        /// The storage name to store information about user preferences.
+        static let preferences = "PreferencesStorage"
     }
 }